Fixed triangular/quadratic index bug

+++ include/util.hpp
@@ -65,11 +65,11 @@ unsigned long long combination(unsigned int n, unsigned int k)
     return result;
 }
 
-unsigned int triangular_index(unsigned int n, unsigned i, unsigned j) {
+unsigned long long triangular_index(unsigned long long n, unsigned long long i, unsigned long long j) {
     return (n * (n - 1) / 2) - (n - i) * ((n - i) - 1) / 2 + j - i - 1;
 }
 
-unsigned int quadratic_index(unsigned int n, unsigned i, unsigned j) {
+unsigned long long quadratic_index(unsigned long long n, unsigned long long i, unsigned long long j) {
     return i * n + j;
 }
